rename src/components/text to src/components/typography
